+++ tests/integration/test_cli.py
@@ -17,7 +17,7 @@ import random
 import shutil
 
 import botocore.session
-from awscli.testutils import unittest, aws
+from awscli.testutils import unittest, aws, BaseS3CLICommand
 from awscli.clidriver import create_clidriver
 
 
@@ -73,7 +73,7 @@ class TestBasicCommandFunctionality(unittest.TestCase):
     def put_object(self, bucket, key, content, extra_args=None):
         session = botocore.session.get_session()
         client = session.create_client('s3', 'us-east-1')
-        response = client.create_bucket(Bucket=bucket)
+        client.create_bucket(Bucket=bucket)
         time.sleep(5)
         self.addCleanup(client.delete_bucket, Bucket=bucket)
         call_args = {
@@ -82,7 +82,7 @@ class TestBasicCommandFunctionality(unittest.TestCase):
         }
         if extra_args is not None:
             call_args.update(extra_args)
-        response = client.put_object(**call_args)
+        client.put_object(**call_args)
         self.addCleanup(client.delete_object, Bucket=bucket, Key=key)
 
     def test_ec2_describe_instances(self):
@@ -97,7 +97,8 @@ class TestBasicCommandFunctionality(unittest.TestCase):
         p = aws('help')
         self.assertEqual(p.rc, 0)
         self.assertIn('AWS', p.stdout)
-        self.assertRegexpMatches(p.stdout, 'The\s+AWS\s+Command\s+Line\s+Interface')
+        self.assertRegexpMatches(
+            p.stdout, 'The\s+AWS\s+Command\s+Line\s+Interface')
 
     def test_service_help_output(self):
         p = aws('ec2 help')
@@ -231,6 +232,9 @@ class TestBasicCommandFunctionality(unittest.TestCase):
         p = aws('s3api list-objects --bucket %s --no-paginate' % bucket_name)
         self.assertEqual(p.rc, 0, p.stdout + p.stderr)
 
+        p = aws('s3api list-objects --bucket %s' % bucket_name)
+        self.assertEqual(p.rc, 0, p.stdout + p.stderr)
+
     def test_top_level_options_debug(self):
         p = aws('ec2 describe-instances --debug')
         self.assertEqual(p.rc, 0)
@@ -320,6 +324,16 @@ class TestBasicCommandFunctionality(unittest.TestCase):
         p = aws('iam list-users', env_vars=base_env_vars)
         self.assertEqual(p.rc, 0)
 
+    def test_error_msg_with_no_region_configured(self):
+        environ = os.environ.copy()
+        try:
+            del environ['AWS_DEFAULT_REGION']
+        except KeyError:
+            pass
+        environ['AWS_CONFIG_FILE'] = 'nowhere-foo'
+        p = aws('ec2 describe-instances', env_vars=environ)
+        self.assertIn('must specify a region', p.stderr)
+
 
 class TestCommandLineage(unittest.TestCase):
     def setUp(self):
@@ -336,7 +350,7 @@ class TestCommandLineage(unittest.TestCase):
         actual_lineage_names = []
         for cmd in command.lineage:
             actual_lineage_names.append(cmd.name)
-        
+
         # Assert the actual names of each command in a lineage is as expected.
         self.assertEqual(actual_lineage_names, ref_lineage_names)
 
@@ -374,6 +388,92 @@ class TestCommandLineage(unittest.TestCase):
     def test_wait_commands(self):
         self.assert_lineage_names(['ec2', 'wait'])
         self.assert_lineage_names(['ec2', 'wait', 'instance-running'])
+
+
+# We're using BaseS3CLICommand because we need a service to use
+# for testing the global arguments.  We're picking S3 here because
+# the BaseS3CLICommand has a lot of utility functions that help
+# with this.
+class TestGlobalArgs(BaseS3CLICommand):
+
+    def test_endpoint_url(self):
+        p = aws('s3api list-objects --bucket dnscompat '
+                '--endpoint-url http://localhost:51515 '
+                '--debug')
+        debug_logs = p.stderr
+        original_hostname = 'dnscompat.s3.amazonaws.com'
+        expected = 'localhost'
+        self.assertNotIn(original_hostname, debug_logs,
+                         '--endpoint-url is being ignored.')
+        self.assertIn(expected, debug_logs)
+
+    def test_no_pagination(self):
+        bucket_name = self.create_bucket()
+        self.put_object(bucket_name, 'foo.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo2.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo3.txt', contents=b'bar')
+        p = aws('s3api list-objects --bucket %s '
+                '--no-paginate --output json' % bucket_name)
+        # A really simple way to check that --no-paginate was
+        # honored is to see if we have all the mirrored input
+        # arguments in the response json.  These normally aren't
+        # present when the response is paginated.
+        self.assert_no_errors(p)
+        response_json = p.json
+        self.assertIn('IsTruncated', response_json)
+        self.assertIn('Name', response_json)
+
+    def test_no_paginate_and_original_args(self):
+        bucket_name = self.create_bucket()
+        self.put_object(bucket_name, 'foo.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo2.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo3.txt', contents=b'bar')
+        p = aws('s3api list-objects --bucket %s '
+                '--max-keys 1 --no-paginate --output json' % bucket_name)
+        self.assert_no_errors(p)
+        response_json = p.json
+        self.assertEqual(len(response_json['Contents']), 1)
+
+    def test_max_items(self):
+        bucket_name = self.create_bucket()
+        self.put_object(bucket_name, 'foo.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo2.txt', contents=b'bar')
+        self.put_object(bucket_name, 'foo3.txt', contents=b'bar')
+        p = aws('s3api list-objects --bucket %s '
+                '--max-items 1 --output json' % bucket_name)
+        self.assert_no_errors(p)
+        response_json = p.json
+        self.assertEqual(len(response_json['Contents']), 1)
+
+    def test_query(self):
+        bucket_name = self.create_bucket()
+        self.put_object(bucket_name, 'foo.txt', contents=b'bar')
+        p = aws('s3api list-objects --bucket %s '
+                '--query Contents[].Key --output json' % bucket_name)
+        self.assert_no_errors(p)
+        response_json = p.json
+        self.assertEqual(response_json, ['foo.txt'])
+
+    def test_no_sign_requests(self):
+        bucket_name = self.create_bucket()
+        self.put_object(bucket_name, 'public', contents=b'bar',
+                        extra_args={'ACL': 'public-read'})
+        self.put_object(bucket_name, 'private', contents=b'bar')
+        env = os.environ.copy()
+        # Set the env vars to bad values so if we do actually
+        # try to sign the request, we'll get an auth error.
+        env['AWS_ACCESS_KEY_ID'] = 'foo'
+        env['AWS_SECRET_ACCESS_KEY'] = 'bar'
+        p = aws('s3api head-object --bucket %s --key public --no-sign-request'
+                % bucket_name, env_vars=env)
+        self.assert_no_errors(p)
+        self.assertIn('ETag', p.json)
+
+        # Should fail because we're not signing the request but the object is
+        # private.
+        p = aws('s3api head-object --bucket %s --key private --no-sign-request'
+                % bucket_name, env_vars=env)
+        self.assertEqual(p.rc, 255)
 
 
 if __name__ == '__main__':